Complete splitting files in the project.
