+++ pytest_html_reporter/__init__.py
@@ -1,1 +1,1 @@
-from .plugin import screenshot as attach
+from .util import screenshot as attach
